// app/data/pipeline/types.ts
export interface RawWeaponProfile {
    datasheetId: string;
    line: string;
    lineInWargear: string;
    dice: string;
    name: string;
    description: string;
    range: string;
    type: string;
    a: string;
    bsWs: string;
    s: string;
    ap: string;
    d: string;
}

export interface RawWeapon {
    id: string;
    datasheetId: string;
    line: string;
    name: string;
    type: string;
    profiles: RawWeaponProfile[];
}

export interface RawOption {
    datasheetId: string;
    line: string;
    button: string;
    description: string;
}